refactor: move attrs tests into its own namespace
