+++ tests/unit/test_contrib/test_attrs/test_schema_spec_generation.py
@@ -1,10 +1,8 @@
 from typing import Dict, List, Optional
 
 import attrs
-from attrs import define
 
 from litestar import post
-from litestar.status_codes import HTTP_201_CREATED
 from litestar.testing import create_test_client
 from tests.models import DataclassPet
 
@@ -51,20 +49,3 @@ def test_spec_generation() -> None:
             "required": ["complex", "first_name", "id", "last_name"],
             "title": "Person",
         }
-
-
-def test_parse_attrs_data_in_signature() -> None:
-    @define(slots=True, frozen=True)
-    class AttrsUser:
-        name: str
-        email: str
-
-    @post("/")
-    async def attrs_data(data: AttrsUser) -> AttrsUser:
-        return data
-
-    with create_test_client([attrs_data]) as client:
-        response = client.post("/", json={"name": "foo", "email": "e@example.com"})
-        assert response.status_code == HTTP_201_CREATED
-        assert response.json().get("name") == "foo"
-        assert response.json().get("email") == "e@example.com"
